server/telegram-bot/requirements.txt
python-telegram-bot~=22.5
aiokafka~=0.12.0
pydantic-settings~=2.11.0
pydantic~=2.12.3